+++ e2e/scripts/st_image.py
@@ -15,5 +15,18 @@
 import streamlit as st
 import numpy as np
 
+st.set_option("deprecation.showImageFormat", True)
 img = np.repeat(0, 10000).reshape(100, 100)
-st.image(img, caption="Black Square")
+st.image(img, caption="Black Square with deprecated format", format="JPEG", width=100)
+
+st.set_option("deprecation.showImageFormat", False)
+st.image(img, caption="Black Square with deprecated format", format="JPEG", width=100)
+
+st.image(img, caption="Black Square as JPEG", output_format="JPEG", width=100)
+
+st.image(img, caption="Black Square as PNG", output_format="PNG", width=100)
+
+st.image(img, caption="Black Square with no output format specified", width=100)
+
+transparent_img = np.zeros((100, 100, 4), dtype=np.uint8)
+st.image(transparent_img, caption="Transparent Black Square", width=100)