+++ frontend/src/components/elements/PlotlyChart/PlotlyChart.test.tsx
@@ -15,8 +15,6 @@
  * limitations under the License.
  */
 
-jest.mock("react-plotly.js", () => jest.fn())
-
 import React from "react"
 import { shallow } from "enzyme"
 import { fromJS } from "immutable"
@@ -25,9 +23,14 @@ import Plot from "react-plotly.js"
 import mock from "./mock"
 import { PropsWithHeight, DEFAULT_HEIGHT } from "./PlotlyChart"
 
-const PlotlyChart = require("./PlotlyChart").PlotlyChart
+jest.mock("react-plotly.js", () => jest.fn())
 
-const getProps = (elementProps: object = {}): PropsWithHeight => ({
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const { PlotlyChart } = require("./PlotlyChart")
+
+const getProps = (
+  elementProps: Record<string, unknown> = {}
+): PropsWithHeight => ({
   element: fromJS({
     ...mock,
     ...elementProps,