+++ frontend/src/components/widgets/Multiselect/Multiselect.test.tsx
@@ -21,14 +21,14 @@ import { fromJS } from "immutable"
 import { multiSelectOverrides } from "lib/widgetTheme"
 import { WidgetStateManager } from "lib/WidgetStateManager"
 
-import Multiselect, { Props } from "./Multiselect"
 import { Select as UISelect, TYPE } from "baseui/select"
+import Multiselect, { Props } from "./Multiselect"
 
 jest.mock("lib/WidgetStateManager")
 
 const sendBackMsg = jest.fn()
 
-const getProps = (elementProps: object = {}): Props => ({
+const getProps = (elementProps: Record<string, unknown> = {}): Props => ({
   element: fromJS({
     id: 1,
     label: "Label",