+++ frontend/src/lib/MetricsManager.test.ts
@@ -15,6 +15,8 @@
  * limitations under the License.
  */
 
+// Disable Typescript checking, since mm.track and identify have private scope
+// @ts-nocheck
 import { SessionInfo } from "lib/SessionInfo"
 import { getMetricsManagerForTest } from "lib/MetricsManagerTestUtils"
 
@@ -32,7 +34,8 @@ beforeEach(() => {
 })
 
 afterEach(() => {
-  SessionInfo["singleton"] = undefined
+  SessionInfo.singleton = undefined
+  window.analytics = undefined
 })
 
 test("does not track while uninitialized", () => {
@@ -42,8 +45,8 @@ test("does not track while uninitialized", () => {
   mm.enqueue("ev2", { data2: 12 })
   mm.enqueue("ev3", { data3: 13 })
 
-  expect(mm["track"].mock.calls.length).toBe(0)
-  expect(mm["identify"].mock.calls.length).toBe(0)
+  expect(mm.track.mock.calls.length).toBe(0)
+  expect(mm.identify.mock.calls.length).toBe(0)
 })
 
 test("does not track when initialized with gatherUsageStats=false", () => {
@@ -54,8 +57,25 @@ test("does not track when initialized with gatherUsageStats=false", () => {
   mm.enqueue("ev2", { data2: 12 })
   mm.enqueue("ev3", { data3: 13 })
 
-  expect(mm["track"].mock.calls.length).toBe(0)
-  expect(mm["identify"].mock.calls.length).toBe(0)
+  expect(mm.track.mock.calls.length).toBe(0)
+  expect(mm.identify.mock.calls.length).toBe(0)
+})
+
+test("does not initialize Segment analytics when gatherUsageStats=false", () => {
+  const mm = getMetricsManagerForTest()
+  expect(window.analytics).toBeUndefined()
+  mm.initialize({ gatherUsageStats: false })
+  expect(window.analytics).toBeUndefined()
+})
+
+test("initializes Segment analytics when gatherUsageStats=true", () => {
+  const mm = getMetricsManagerForTest()
+  expect(window.analytics).toBeUndefined()
+  mm.initialize({ gatherUsageStats: true })
+  expect(window.analytics).toBeDefined()
+  expect(window.analytics.invoked).toBe(true)
+  expect(window.analytics.methods).toHaveLength(20)
+  expect(window.analytics.load).toBeDefined()
 })
 
 test("enqueues events before initialization", () => {
@@ -65,25 +85,25 @@ test("enqueues events before initialization", () => {
   mm.enqueue("ev2", { data2: 12 })
   mm.enqueue("ev3", { data3: 13 })
 
-  expect(mm["track"].mock.calls.length).toBe(0)
+  expect(mm.track.mock.calls.length).toBe(0)
 
   mm.initialize({ gatherUsageStats: true })
 
-  expect(mm["track"].mock.calls.length).toBe(3)
-  expect(mm["identify"].mock.calls.length).toBe(1)
+  expect(mm.track.mock.calls.length).toBe(3)
+  expect(mm.identify.mock.calls.length).toBe(1)
 })
 
 test("tracks events immediately after initialized", () => {
   const mm = getMetricsManagerForTest()
   mm.initialize({ gatherUsageStats: true })
 
-  expect(mm["track"].mock.calls.length).toBe(0)
+  expect(mm.track.mock.calls.length).toBe(0)
   mm.enqueue("ev1", { data1: 11 })
-  expect(mm["track"].mock.calls.length).toBe(1)
+  expect(mm.track.mock.calls.length).toBe(1)
   mm.enqueue("ev2", { data2: 12 })
-  expect(mm["track"].mock.calls.length).toBe(2)
+  expect(mm.track.mock.calls.length).toBe(2)
   mm.enqueue("ev3", { data3: 13 })
-  expect(mm["track"].mock.calls.length).toBe(3)
+  expect(mm.track.mock.calls.length).toBe(3)
 })
 
 test("increments deltas", () => {
@@ -95,7 +115,7 @@ test("increments deltas", () => {
   mm.incrementDeltaCounter("foo")
   mm.incrementDeltaCounter("bar")
 
-  const counter = mm.getDeltaCounter()
+  const counter = mm.getAndResetDeltaCounter()
 
   expect(counter.foo).toBe(3)
   expect(counter.bar).toBe(2)
@@ -112,7 +132,7 @@ test("clears deltas", () => {
   mm.incrementDeltaCounter("bar")
 
   mm.clearDeltaCounter()
-  const counter = mm.getDeltaCounter()
+  const counter = mm.getAndResetDeltaCounter()
 
   expect(Object.keys(counter).length).toBe(0)
 })
@@ -126,8 +146,8 @@ test("clears deltas automatically on read", () => {
   mm.incrementDeltaCounter("foo")
   mm.incrementDeltaCounter("bar")
 
-  const counter1 = mm.getDeltaCounter()
-  const counter2 = mm.getDeltaCounter()
+  const counter1 = mm.getAndResetDeltaCounter()
+  const counter2 = mm.getAndResetDeltaCounter()
 
   expect(Object.keys(counter1).length).toBe(2)
   expect(Object.keys(counter2).length).toBe(0)