+++ frontend/src/lib/MetricsManager.ts
@@ -15,13 +15,16 @@
  * limitations under the License.
  */
 
-import { IS_DEV_ENV, IS_SHARED_REPORT } from "./baseconsts"
 import { SessionInfo } from "lib/SessionInfo"
+import { IS_DEV_ENV, IS_SHARED_REPORT } from "./baseconsts"
 import { logAlways } from "./log"
+import { initializeSegment } from "./Segment"
 
 /**
- * The analytics is the Segment.io object. It comes from index.html.
- */
+ * The analytics is the Segment.io object. It is initialized in Segment.ts
+ * It is loaded with global scope (window.analytics) to integrate with the segment.io api
+ * @global
+ * */
 declare const analytics: any
 
 /**
@@ -32,7 +35,15 @@ interface DeltaCounter {
   [name: string]: number
 }
 
-type Event = [string, object]
+/**
+ * A mapping of [component instance name] -> [count] which is used to upload
+ * custom component stats when the app is idle.
+ */
+interface CustomComponentCounter {
+  [name: string]: number
+}
+
+type Event = [string, Record<string, unknown>]
 
 export class MetricsManager {
   private initialized = false
@@ -55,6 +66,20 @@ export class MetricsManager {
   private pendingDeltaCounter: DeltaCounter = {}
 
   /**
+   * Object used to count the number of custom instance names seen in a given report.
+   * Maps type of custom instance name (string) to count (number).
+   */
+  private pendingCustomComponentCounter: CustomComponentCounter = {}
+
+  /**
+   * Report hash uniquely identifies "projects" so we can tell
+   * how many projects are being created with Streamlit while still keeping
+   * possibly-sensitive info like the scriptPath outside of our metrics
+   * services.
+   */
+  private reportHash = "Not initialized"
+
+  /**
    * Singleton MetricsManager object. The reason we're using a singleton here
    * instead of just exporting a module-level instance is so we can easily
    * override it in tests.
@@ -70,10 +95,12 @@ export class MetricsManager {
     this.actuallySendMetrics = gatherUsageStats
 
     if (this.actuallySendMetrics || IS_SHARED_REPORT) {
+      // Segment will not initialize if this is rendered with SSR
+      initializeSegment()
       // Only record the user's email if they entered a non-empty one.
       const userTraits: any = {}
       if (SessionInfo.current.authorEmail !== "") {
-        userTraits["authoremail"] = SessionInfo.current.authorEmail
+        userTraits.authoremail = SessionInfo.current.authorEmail
       }
       this.identify(SessionInfo.current.installationId, userTraits)
       this.sendPendingEvents()
@@ -82,7 +109,7 @@ export class MetricsManager {
     logAlways("Gather usage stats: ", this.actuallySendMetrics)
   }
 
-  public enqueue(evName: string, evData: object = {}): void {
+  public enqueue(evName: string, evData: Record<string, unknown> = {}): void {
     if (!this.initialized) {
       this.pendingEvents.push([evName, evData])
       return
@@ -107,15 +134,42 @@ export class MetricsManager {
     }
   }
 
-  public getDeltaCounter(): DeltaCounter {
+  public getAndResetDeltaCounter(): DeltaCounter {
     const deltaCounter = this.pendingDeltaCounter
-    this.pendingDeltaCounter = {}
+    this.clearDeltaCounter()
     return deltaCounter
   }
 
-  private send(evName: string, evData: object = {}): void {
+  public clearCustomComponentCounter(): void {
+    this.pendingCustomComponentCounter = {}
+  }
+
+  public incrementCustomComponentCounter(customInstanceName: string): void {
+    if (this.pendingCustomComponentCounter[customInstanceName] == null) {
+      this.pendingCustomComponentCounter[customInstanceName] = 1
+    } else {
+      this.pendingCustomComponentCounter[customInstanceName]++
+    }
+  }
+
+  public getAndResetCustomComponentCounter(): CustomComponentCounter {
+    const customComponentCounter = this.pendingCustomComponentCounter
+    this.clearCustomComponentCounter()
+    return customComponentCounter
+  }
+
+  // Report hash gets set when update report happens.
+  // This means that it will be attached to most, but not all, metrics events.
+  // The viewReport and createReport events are sent before updateReport happens,
+  // so they will not include the reportHash.
+  public setReportHash = (reportHash: string): void => {
+    this.reportHash = reportHash
+  }
+
+  private send(evName: string, evData: Record<string, unknown> = {}): void {
     const data = {
       ...evData,
+      reportHash: this.reportHash,
       dev: IS_DEV_ENV,
       source: "browser",
       streamlitVersion: SessionInfo.current.streamlitVersion,
@@ -139,12 +193,13 @@ export class MetricsManager {
   }
 
   // Wrap analytics methods for mocking:
-
-  private identify(id: string, data: object): void {
+  // eslint-disable-next-line class-methods-use-this
+  private identify(id: string, data: Record<string, unknown>): void {
     analytics.identify(id, data)
   }
 
-  private track(evName: string, data: object): void {
+  // eslint-disable-next-line class-methods-use-this
+  private track(evName: string, data: Record<string, unknown>): void {
     analytics.track(evName, data)
   }
 }