+++ lib/tests/streamlit/config_test.py
@@ -13,20 +13,18 @@
 # limitations under the License.
 
 """Config System Unittest."""
+from unittest.mock import MagicMock, mock_open, patch
 import copy
 import os
 import textwrap
 import unittest
 
 import pytest
-from mock import MagicMock
-from mock import mock_open
-from mock import patch
 from parameterized import parameterized
 
 from streamlit import config
 from streamlit import env_util
-from streamlit.ConfigOption import ConfigOption
+from streamlit.config_option import ConfigOption
 
 SECTION_DESCRIPTIONS = copy.deepcopy(config._section_descriptions)
 CONFIG_OPTIONS = copy.deepcopy(config._config_options)
@@ -257,6 +255,7 @@ class ConfigTest(unittest.TestCase):
                 "_test",
                 "browser",
                 "client",
+                "deprecation",
                 "global",
                 "mapbox",
                 "runner",
@@ -275,6 +274,8 @@ class ConfigTest(unittest.TestCase):
                 "browser.serverPort",
                 "client.caching",
                 "client.displayEnabled",
+                "deprecation.showfileUploaderEncoding",
+                "deprecation.showImageFormat",
                 "global.developmentMode",
                 "global.disableWatchdogWarning",
                 "global.logLevel",
@@ -285,7 +286,6 @@ class ConfigTest(unittest.TestCase):
                 "global.showWarningOnDirectExecution",
                 "global.suppressDeprecationWarnings",
                 "global.unitTest",
-                "global.useNode",
                 "runner.magicEnabled",
                 "runner.installTracer",
                 "runner.fixMatplotlib",